+++ tests/test_comments.py
@@ -29,5 +29,20 @@ def test_incorrect_fixture(absolute_path):
     assert stdout.count(b'E800') == 1
     assert b'-    # property_name = 1' in stdout
     assert b'-    # typed_property: int = 10' in stdout
+
+
+def test_incorrect_fixture_aggressive(absolute_path):
+    """End-to-End test to check that incorrect code raises warning."""
+    filename = absolute_path('fixtures', 'incorrect.py')
+    process = subprocess.Popen(
+        ['flake8', '--eradicate-aggressive', '--select', 'E8', filename],
+        stdout=subprocess.PIPE,
+        stderr=subprocess.PIPE,
+    )
+    stdout, _ = process.communicate()
+
+    assert stdout.count(b'E800') == 1
+    assert b'-    # property_name = 1' in stdout
+    assert b'-    # typed_property: int = 10' in stdout
     assert b'-# def function_name():' in stdout
     assert b'-# class CommentedClass(object):' in stdout